Refactor lsif input parsing.
