+++ cmd/precise-code-intel-worker/internal/correlation/lsif/jsonlines/reader_test.go
@@ -1,0 +1,3 @@
+package jsonlines
+
+// TODO(efritz) - write tests
