+++ cmd/precise-code-intel-worker/internal/correlation/lsif/reader.go
@@ -1,0 +1,13 @@
+package lsif
+
+import (
+	"context"
+	"io"
+)
+
+type Pair struct {
+	Element Element
+	Err     error
+}
+
+type Reader func(ctx context.Context, r io.Reader) <-chan Pair
